% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{extract}
\alias{extract}
\title{Get zonal statistics from raster}
\usage{
extract(r, geoms)
}
\arguments{
\item{r}{Raster or RasterBrick object}

\item{shp}{WKB object, e.g. \code{sf::st_as_binary(sf::st_geometry(basins), EWKB=TRUE)}}
}
\description{
Get zonal statistics from raster
}
