% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2-package.R
\docType{package}
\name{extract2-package}
\alias{extract2}
\alias{extract2-package}
\title{extract2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
