% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{overlap}
\alias{overlap}
\title{Get the detailed information of overlapped grids}
\usage{
overlap(r, shp)
}
\arguments{
\item{r}{Raster or RasterBrick object}

\item{shp}{WKB object, e.g. \code{sf::st_as_binary(sf::st_geometry(basins), EWKB=TRUE)}}
}
\value{
geoms
\itemize{
\item \code{weights}: fraction percentage
\item \code{area}: area in km^2
\item \code{row},\code{col}: begining \code{row} and \code{col} of the overlapped region
\item \code{nrow},\code{ncol}: number of the rows and columns of the overlapped region
}
}
\description{
Get the detailed information of overlapped grids
}
