% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract2.R
\name{sf_func}
\alias{sf_func}
\alias{sf_mean}
\alias{sf_sum}
\alias{sf_mean_weighted}
\alias{sf_sum_weighted}
\title{functions of zonal statistics}
\usage{
sf_mean(vals, cov_fracs, .fun_sum = colSums, ...)

sf_sum(vals, cov_fracs, .fun_sum = colSums, ...)

sf_mean_weighted(vals, cov_fracs, weights, .fun_sum = colSums, ...)

sf_sum_weighted(vals, cov_fracs, weights, .fun_sum = colSums, ...)
}
\description{
functions of zonal statistics
}
